Minor type safety fixes

+++ src/core/util/TimeUtils.cpp
@@ -19,7 +19,7 @@ Timer::~Timer() {
     if (performanceTracker != nullptr) {
         performanceTracker->recordValue(name, startNs, endNs);
     } else {
-        long long duration = endNs - startNs;
+        int64_t duration = endNs - startNs;
         const double conversionFactor = 0.000001;
         double ms = static_cast<double>(duration) * conversionFactor;
         std::cout << name.c_str() << ": " << ms << "ms\n";
